Update servant_names.json
Short answer: No, we will not change that.

Long answer: The romanized names in mahoyo (at least, most of them, I'll get to that) are romanized in a system called kunrei-shiki, which is the system preferred by the Japanese government, which might make you think it should be defaulted to, but it's far from the most popular or most commonly used system, that being hepburn. The fact that the names in mahoyo are romanized like this instead of with hepburn is most likely a decision from the Japanese higher-ups, rather than anything on AoA's side, but even that remains very inconsistent, with Tsukihime stuff using hepburn, for example. 
It's very likely that just as how Altria is only what it is because Nasu got the name of the cigarette brand from google way back in the day, "Sizuki" is what he had written in his original mahoyo draft, and he's simply sticking to it.

The issue, of course, is that the Japanese are notoriously bad with english and I would say most JP-speakers agree that kunrei-shiki only sort of exaggerates this idea.
Using "Sizuki Soujyuro" should mean that you also want to use "Arkaid Brynsted" or "Ilyasfeel von Einzbren", for example. These are all official name spellings, and they are all awful.

Adhering to the whims of the Japanese executives who decide on this kind of stuff only makes sense insofar as they have any idea what they're actually working with, a category to which english spelling very rarely falls.